# swagger.yaml
openapi: '3.0.0'
info:
  title: Author's Haven
  version: '1.0.0'
  description: A Social platform for the creative at heart.
  contact:
    name: Andela Cohort 43 Team Neon
servers:
  - url: https://neon-ah-staging.herokuapp.com/api/v1/
...
components:
  schemas:
    Error:
      properties:
        message:
          description: Error message
          type: string
...
  responses:
      Success:
        description: Success
...
      ErrorBadRequest:
        description: Bad Request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Error'
      ErrorUnauthorized:
        description: Unauthorized
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Error'
...
      ErrorInternalServer:
        description: Internal Server Error
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Error'   
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT 
paths:
...
  /upgrade/user/{userName}:
    put:
      tags:
        - Users
      summary: This method upgrades a user to admin. It accessible only to the superadmin.
      parameters:
        - name: userName
          in: path
          description: ID of the User to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '500':
          $ref: '#/components/responses/ErrorInternalServer'
  /users/{userName}:
    get:
      tags:
        - Users
      summary: This method upgrades a user to admin. It accessible only to the superadmin.
      parameters:
        - name: userName
          in: path
          description: ID of the User to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '500':
          $ref: '#/components/responses/ErrorInternalServer'
...